General statistic

+++ Models/Models/Statistic/StatisticGeneralDo.cs
@@ -7,7 +7,7 @@ namespace Models.Models.Statistic
     public class StatisticGeneralDo
     {
         public double meanTime { get; set; }
-        public int meanCount { get; set; }
+        public double meanCount { get; set; }
         public IEnumerable<string> chartLabels { get; set; }
         public IEnumerable<tableDataa> dataForTable { get; set; }
